chore: refactor cover service
