--- services/cover/src/service/cover.rs
@@ -1,163 +1,0 @@
-use crate::common::types::{CallerId, CanisterId, ProviderId, ReqId};
-use crate::service::guard::{is_cover_owner, is_valid_provider};
-use crate::service::types::{
-    AddProvider, AddVerification, CreateRequest, Error, Progress, Provider, ProviderInfo, Request,
-    SubmitVerification, UpdateProgress, UpdateProvider, UpdateVerification, Verification,
-};
-
-use super::{
-    get_progress_store, get_progress_store_mut, get_provider_store, get_provider_store_mut,
-    get_request_store, get_request_store_mut, get_verification_store, get_verification_store_mut,
-};
-
-pub fn get_request_by_id(request_id: ReqId) -> Option<&'static Request> {
-    get_request_store().get_request_by_id(request_id)
-}
-
-pub fn get_all_requests() -> Vec<&'static Request> {
-    get_request_store().get_all_requests()
-}
-
-pub fn get_progress_by_request_id(request_id: ReqId) -> Option<&'static Progress> {
-    get_progress_store().get_progress_by_request_id(request_id)
-}
-
-pub fn get_progresses_by_canister_id(canister_id: CanisterId) -> Vec<&'static Progress> {
-    get_progress_store().get_progresses_by_canister_id(canister_id)
-}
-
-pub fn get_all_progresses() -> Vec<&'static Progress> {
-    get_progress_store().get_all_progresses()
-}
-
-pub fn get_verification_by_canister_id(canister_id: &CanisterId) -> Option<&'static Verification> {
-    get_verification_store().get_verification_by_canister_id(canister_id)
-}
-
-pub fn get_all_verifications() -> Vec<&'static Verification> {
-    get_verification_store().get_all_verifications()
-}
-
-pub fn consume_requests(provider_info: ProviderInfo) -> Result<Vec<&'static Request>, Error> {
-    // TODO: check allow list
-    get_request_store_mut()
-        .consume_requests(provider_info)
-        .and_then(|requests| {
-            for request in requests.iter() {
-                get_progress_store_mut().init_progress(request.request_id, request.canister_id)?;
-            }
-            Ok(requests)
-        })
-        .map_err(|e| e.into())
-}
-
-pub fn update_progress(update_progress: UpdateProgress) -> Result<(), Error> {
-    // TODO: check progress owner
-    get_progress_store_mut()
-        .update_progress(update_progress)
-        .map_err(|e| e.into())
-}
-
-pub fn add_verification(
-    caller_id: CallerId,
-    add_verification: AddVerification,
-) -> Result<(), Error> {
-    is_valid_provider(&caller_id, || {
-        get_verification_store_mut()
-            .add_verification(caller_id, add_verification)
-            .map_err(|e| e.into())
-    })
-}
-
-pub fn update_verification(
-    caller_id: CallerId,
-    update_verification: UpdateVerification,
-) -> Result<(), Error> {
-    is_valid_provider(&caller_id, || {
-        get_verification_store_mut()
-            .update_verification(caller_id, update_verification)
-            .map_err(|e| e.into())
-    })
-}
-
-pub fn submit_verification(
-    caller_id: CallerId,
-    submit_verification: SubmitVerification,
-) -> Result<(), Error> {
-    is_valid_provider(&caller_id, || {
-        match get_verification_store().verification_exists(&submit_verification.canister_id) {
-            true => get_verification_store_mut()
-                .update_verification(
-                    caller_id,
-                    UpdateVerification {
-                        canister_id: submit_verification.canister_id,
-                        git_sha: submit_verification.git_sha,
-                        git_ref: submit_verification.git_ref,
-                        git_repo: submit_verification.git_repo,
-                        wasm_checksum: submit_verification.wasm_checksum,
-                        build_log_url: submit_verification.build_log_url,
-                        source_snapshot_url: submit_verification.source_snapshot_url,
-                    },
-                )
-                .map_err(|e| e.into()),
-            false => get_verification_store_mut()
-                .add_verification(
-                    caller_id,
-                    AddVerification {
-                        canister_id: submit_verification.canister_id,
-                        git_sha: submit_verification.git_sha,
-                        git_ref: submit_verification.git_ref,
-                        git_repo: submit_verification.git_repo,
-                        wasm_checksum: submit_verification.wasm_checksum,
-                        build_log_url: submit_verification.build_log_url,
-                        source_snapshot_url: submit_verification.source_snapshot_url,
-                    },
-                )
-                .map_err(|e| e.into()),
-        }
-    })
-}
-
-pub fn create_request(caller_id: CallerId, create_request: CreateRequest) -> Result<(), Error> {
-    // TODO: handle canister's owner properly
-    get_request_store_mut().create_request(caller_id, create_request);
-    Ok(())
-}
-
-pub async fn add_provider(caller_id: CallerId, add_provider: AddProvider) -> Result<(), Error> {
-    is_cover_owner(&caller_id, || {
-        get_provider_store_mut()
-            .add_provider(caller_id, add_provider)
-            .map_err(|e| e.into())
-    })
-    .await
-}
-
-pub async fn update_provider(
-    caller_id: CallerId,
-    update_provider: UpdateProvider,
-) -> Result<(), Error> {
-    is_cover_owner(&caller_id, || {
-        get_provider_store_mut()
-            .update_provider(caller_id, update_provider)
-            .map_err(|e| e.into())
-    })
-    .await
-}
-
-pub async fn delete_provider(caller_id: &CallerId, provider_id: &ProviderId) -> Result<(), Error> {
-    is_cover_owner(caller_id, || {
-        get_provider_store_mut()
-            .delete_provider(provider_id)
-            .map_err(|e| e.into())
-    })
-    .await
-}
-
-pub fn get_provider_by_id(provider_id: &ProviderId) -> Option<&'static Provider> {
-    get_provider_store().get_provider_by_id(provider_id)
-}
-
-pub fn get_all_providers() -> Vec<&'static Provider> {
-    get_provider_store().get_all_providers()
-}
